Refactor component paths and update registry config
Renamed component files to use lowercase paths for consistency and updated all imports accordingly. Modified registry.json to reflect new paths and improved registryDependencies to use full URLs. Cleaned up Button component props and updated related stories and tests to match new import paths.
