+++ src/components/ui/accordion.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as AccordionPrimitive from "@radix-ui/react-accordion"
 import { cn } from "@/lib/utils"
-import { Icon } from "./Icon/Icon"
+import { Icon } from "@/components/ui/icon"
 
 const Accordion = AccordionPrimitive.Root
