+++ src/components/ui/dialog.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as DialogPrimitive from "@radix-ui/react-dialog"
 import { cn } from "@/lib/utils"
-import { Icon } from "./Icon/Icon"
+import { Icon } from "./icon"
 
 const Dialog = DialogPrimitive.Root
