+++ src/components/ui/pagination.tsx
@@ -2,7 +2,7 @@
 
 import * as React from "react"
 import { cn } from "@/lib/utils"
-import { Icon } from "./Icon/Icon"
+import { Icon } from "./icon"
 
 const Pagination = ({ className, ...props }: React.ComponentProps<"nav">) => (
   <nav
