+++ src/test/icon.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 import { render, screen } from '@testing-library/react'
-import { Icon, SearchIcon, HomeIcon, PersonIcon } from './Icon'
+import { Icon, SearchIcon, HomeIcon, PersonIcon } from '@/components/ui/icon'
 
 describe('Icon', () => {
   describe('Basic rendering', () => {
